% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{convertIdx}
\alias{convertIdx}
\title{Converti charactere en numeric}
\usage{
convertIdx(i, names, n = length(names), invert = FALSE)
}
\arguments{
\item{invert}{}
}
\description{
Converti charactere en numeric
}
