% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JS_Text.R
\name{JS_TEXT}
\alias{JS_TEXT}
\title{Title}
\usage{
JS_TEXT(
  data,
  Stat = Stat,
  Note = Note,
  Lecture = Lecture,
  Source = Source,
  Champs = Champs,
  ...
)
}
\arguments{
\item{data}{La base de donnees}

\item{Stat}{Une chaine de caracteres}

\item{Note}{Une chaine de caracteres}

\item{Lecture}{Une chaine de caracteres}

\item{Source}{Une chaine de caracteres}

\item{Champs}{Une chaine de caracteres}

\item{...}{Parametre autre}
}
\value{

}
\description{
Title
}
