% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tableau_Option.R
\name{TABLEAU_OPTION}
\alias{TABLEAU_OPTION}
\title{Option for datatable}
\usage{
TABLEAU_OPTION(
  data = data,
  dom = dom,
  Reduction = Reduction,
  ID = ID,
  Stat = Stat,
  Note = Note,
  Lecture = Lecture,
  Source = Source,
  Champs = Champs,
  ...
)
}
\arguments{
\item{dom}{la chaine de caractère correspondant au dom}

\item{...}{other}
}
\value{

}
\description{
Option for datatable
}
