% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{TABLEAU_TITRE}
\alias{TABLEAU_TITRE}
\title{Title}
\usage{
TABLEAU_TITRE(data, number = NULL, title = NULL, color = color, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
