% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tableau_Convert.R
\name{convertIdx}
\alias{convertIdx}
\title{Convertion caractere en numeric}
\usage{
convertIdx(i, names, n = length(names), invert = FALSE)
}
\arguments{
\item{i}{iteration}

\item{names}{chaine de caractere}

\item{n}{longueur}

\item{invert}{invertion}
}
\value{
une convertion de character en numeric ou l'inverse
}
\description{
Convertion caractere en numeric
}
