% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.R
\name{TABLEAU_OPTION}
\alias{TABLEAU_OPTION}
\title{Title}
\usage{
TABLEAU_OPTION(data, Lecture = Lecture, Source = Source,
  Champs = Champs, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
