% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tableau_Option.R
\name{TABLEAU_OPTION}
\alias{TABLEAU_OPTION}
\title{Title}
\usage{
TABLEAU_OPTION(
  dom = dom,
  Stat = Stat,
  Note = Note,
  Lecture = Lecture,
  Source = Source,
  Champs = Champs,
  ...
)
}
\arguments{
\item{dom}{la chaine de caractère correspondant au dom}

\item{...}{}
}
\value{

}
\description{
Title
}
