% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tableau_Titre.R
\name{TABLEAU_TITRE}
\alias{TABLEAU_TITRE}
\title{Title}
\usage{
TABLEAU_TITRE(data = data, number = NULL, title = NULL, color = color, ...)
}
\arguments{
\item{data}{La base de donnees}

\item{number}{Une chaine de caractere ou un chiffre qui viendra se mettre devant le titre sous forme : "Tableau .... : "}

\item{title}{Une chaine de caractere qui sera le titre}

\item{color}{Chaine de caractere qui peut etre soit au format HSL, HEX, RGB, ou une chaine de caractere qui correspond a la palette de couleur de base de R.}

\item{...}{Parametre autre}
}
\value{

}
\description{
Title
}
